+++ src/Exp.c
@@ -89,18 +89,26 @@ SEXP Call_dexp1(SEXP x, /*  double                          */
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ++i) {
 
-      if (ISNA(rx[ix]) || (rscale[iscale] <= 0.0)) {
-
-	rval[i] = NA_REAL;
+      if (!R_FINITE(rx[ix]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
+	
+	if ((rx[ix] == R_NegInf) || (rx[ix] == R_PosInf)) {
+	  rval[i] = R_NegInf;
+	  if (!INTEGER(logFlag)[0]) {
+	    rval[i] = exp(rval[i]);
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
+	
 	rgrad[i] = NA_REAL;
-
+	
 	if (hessian) {
 	  // row 'sigma'
 	  rhess[i] = NA_REAL;
 	}
-
+	
       } else {
-
+	
 	z = rx[ix] / rscale[iscale];
 	sigma = rscale[iscale];
 
@@ -159,13 +167,20 @@ SEXP Call_dexp1(SEXP x, /*  double                          */
 	 ix = (++ix == nx) ? 0 : ix,
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ++i) {
-
-      if (ISNA(rx[ix]) || (rscale[iscale] <= 0.0)) {
-
-	rval[i] = NA_REAL;
-
+      
+      if (!R_FINITE(rx[ix]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
+	
+       	if ((rx[ix] == R_NegInf) || (rx[ix] == R_PosInf)) {
+	  rval[i] = R_NegInf;
+	  if (!INTEGER(logFlag)[0]) {
+	    rval[i] = exp(rval[i]);
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
+	
       } else {
-
+	
 	z = rx[ix] / rscale[iscale];
 	// Rprintf("%d, %d, %6.3f, sigma = %6.3f\n", i, ishape, xi, rscale[iscale]);
 
@@ -257,21 +272,36 @@ SEXP Call_pexp1(SEXP q,               /*  double                          */
 	// row 'mu'
 	rhess[i] = 0.0;
       }
+      
+      if (!R_FINITE(rq[iq]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
 
-      if (ISNA(rq[iq]) || (rscale[iscale] <= 0.0)) {
+	if (rq[iq] == R_NegInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 0.0;
+	  } else {
+	    rval[i] = 1.0;
+	  }
+	} else if (rq[iq] == R_PosInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 1.0;
+	  } else {
+	    rval[i] = 0.0;
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
 
-	rval[i] = NA_REAL;
 	rgrad[i] = NA_REAL;
-
+	
 	if (hessian) {
 	  rhess[i] = NA_REAL;
 	}
-
+	
       } else {
-
+	
 	z = rq[iq] / rscale[iscale];
 	sigma = rscale[iscale];
-
+	
 	if (z < 0.0) {
 	  S = 1.0;
 	  rval[i] = S;
@@ -315,9 +345,25 @@ SEXP Call_pexp1(SEXP q,               /*  double                          */
 	 iq = (++iq == nq) ? 0 : iq,
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ++i) {
-
-      if (ISNA(rq[iq]) || (rscale[iscale] <= 0.0)) {
-	rval[i] = NA_REAL;
+      
+      if (!R_FINITE(rq[iq]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
+	
+	if (rq[iq] == R_NegInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 0.0;
+	  } else {
+	    rval[i] = 1.0;
+	  }
+	} else if (rq[iq] == R_PosInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 1.0;
+	  } else {
+	    rval[i] = 0.0;
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
+	
       } else {
 	z = rq[iq] / rscale[iscale];
 	if (z < 0.0) {
@@ -399,9 +445,9 @@ SEXP Call_qexp1(SEXP p,               /*  double                          */
 	 ip = (++ip == np) ? 0 : ip,
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ++i) {
-
-      if (ISNA(rp[ip]) || (rscale[iscale] <= 0.0)) {
-
+      
+      if (ISNA(rp[ip]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
+      
 	rval[i] = NA_REAL;
 	rgrad[i] = NA_REAL;
 	rhess[i] = NA_REAL;
@@ -459,7 +505,7 @@ SEXP Call_qexp1(SEXP p,               /*  double                          */
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ++i) {
 
-      if (ISNA(rp[ip]) || (rscale[iscale] <= 0.0)) {
+      if (ISNA(rp[ip]) || !R_FINITE(rscale[iscale]) || (rscale[iscale] <= 0.0)) {
 
 	rval[i] = NA_REAL;