+++ src/GEV.c
@@ -103,9 +103,17 @@ SEXP Call_dGEV(SEXP x,             /*  double                          */
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
       
-      if (ISNA(rx[ix]) || (rscale[iscale] <= 0.0)) {
-	
-	rval[i] = NA_REAL;
+      if (!R_FINITE(rx[ix]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
+
+	if ((rx[ix] == R_NegInf) || (rx[ix] == R_PosInf)) {
+	  rval[i] = R_NegInf;
+	  if (!INTEGER(logFlag)[0]) {
+	    rval[i] = exp(rval[i]);
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
 	
 	rgrad[i] = NA_REAL;
 	rgrad[i + n] = NA_REAL;
@@ -299,11 +307,19 @@ SEXP Call_dGEV(SEXP x,             /*  double                          */
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
-      
-      if (ISNA(rx[ix]) || (rscale[iscale] <= 0.0)) {
+      	    
+      if (!R_FINITE(rx[ix]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
 	
-	rval[i] = NA_REAL;
-		
+	if ((rx[ix] == R_NegInf) || (rx[ix] == R_PosInf)) {
+	  rval[i] = R_NegInf;
+	  if (!INTEGER(logFlag)[0]) {
+	    rval[i] = exp(rval[i]);
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
+	
       } else {
 	
 	z = (rx[ix] - rloc[iloc]) / rscale[iscale];
@@ -410,15 +426,29 @@ SEXP Call_pGEV(SEXP q,               /*  double                          */
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
       
-      if (ISNA(rq[iq]) || (rscale[iscale] <= 0.0)) {
-	
-	rval[i] = NA_REAL;
+      if (!R_FINITE(rq[iq]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
+       
+	if (rq[iq] == R_NegInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 0.0;
+	  } else {
+	    rval[i] = 1.0;
+	  }
+	} else if (rq[iq] == R_PosInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 1.0;
+	  } else {
+	    rval[i] = 0.0;
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
 	
 	rgrad[i] = NA_REAL;
 	rgrad[i + n] = NA_REAL;
 	rgrad[i + 2 * n] = NA_REAL; 
 	
-
       } else {
 	
 	z = (rq[iq] - rloc[iloc]) / rscale[iscale];
@@ -502,10 +532,26 @@ SEXP Call_pGEV(SEXP q,               /*  double                          */
 	   iscale = (++iscale == nscale) ? 0 : iscale,
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
-      
-      if (ISNA(rq[iq]) || (rscale[iscale] <= 0.0)) {
+
+      // a non-finite value of rq can 
+      if (!R_FINITE(rq[iq]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
 	
-	rval[i] = NA_REAL;
+	if (rq[iq] == R_NegInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 0.0;
+	  } else {
+	    rval[i] = 1.0;
+	  }
+	} else if (rq[iq] == R_PosInf) {
+	  if (INTEGER(lowerTailFlag)[0]) {
+	    rval[i] = 1.0;
+	  } else {
+	    rval[i] = 0.0;
+	  }
+	} else {
+	  rval[i] = NA_REAL;
+	}
 		
       } else {
 	
@@ -626,10 +672,11 @@ SEXP Call_qGEV(SEXP p,               /*  double                          */
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
       
-      
-      if (ISNA(rp[ip]) || (rscale[iscale] <= 0.0)) {
+      if (ISNA(rp[ip]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
+	
 	// Rprintf("NA case\n");
-
+	
 	rval[i] = NA_REAL;
 	
 	for (j = 0; j < 3; j++) {
@@ -823,7 +870,8 @@ SEXP Call_qGEV(SEXP p,               /*  double                          */
 	   ishape = (++ishape == nshape) ? 0 : ishape,
 	   ++i) {
       
-      if (ISNA(rp[ip]) || (rscale[iscale] <= 0.0)) {
+      if (ISNA(rp[ip]) || !R_FINITE(rloc[iloc]) || !R_FINITE(rscale[iscale]) ||
+	  !R_FINITE(rshape[ishape]) || (rscale[iscale] <= 0.0)) {
 	
 	rval[i] = NA_REAL;